+++ .github/workflows/ci.yml
@@ -18,6 +18,22 @@ jobs:
       script: 'yarn build && yarn test'
       nodeVersions: '[14,16,18]'
 
+  deno:
+    name: Deno Test
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout Repository
+        uses: actions/checkout@v3
+
+      - name: Setup Deno
+        uses: denoland/setup-deno@v1
+        with:
+          deno-version: v1.x
+
+      - name: Run tests
+        working-directory: ./examples/deno
+        run: deno test ./test.ts
+
   integration:
     name: integration / nodejs v${{ matrix.node-version }}
     runs-on: ubuntu-latest
