+++ examples/deno/index.ts
@@ -1,23 +1,5 @@
-import {
-  createYoga,
-  createSchema,
-} from 'https://cdn.skypack.dev/graphql-yoga@three?dts'
 import { serve } from 'https://deno.land/std@0.157.0/http/server.ts'
-
-const yoga = createYoga({
-  schema: createSchema({
-    typeDefs: /* GraphQL */ `
-      type Query {
-        hello: String!
-      }
-    `,
-    resolvers: {
-      Query: {
-        hello: () => 'Hello Deno!',
-      },
-    },
-  }),
-})
+import { yoga } from './yoga.ts'
 
 serve(yoga, {
   onListen({ hostname, port }) {
